Time: 23 ms (36.43%) | Memory: 14.6 MB (24.89%) - LeetSync

+++ 227-basic-calculator-ii/basic-calculator-ii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int calculate(string s) {
-        long long int curr = 0, ans = 0;
+         int curr = 0, ans = 0;
         s += '+';
         stack<int> stk;
         char op = '+';
